refactor: consolidate OpenAI classes into openai module
- Create ClaudeSwarm::OpenAI module to organize OpenAI-related classes
- Move OpenAIExecutor to OpenAI::Executor
- Move OpenAIChatCompletion to OpenAI::ChatCompletion
- Move OpenAIResponses to OpenAI::Responses
- Update all references throughout codebase
- Fix namespace conflict with ::OpenAI::Client
- Update tests to use new module structure

This improves code organization by grouping related OpenAI functionality
into a dedicated module namespace.
